Move csv_bulk_upload logic to util functions
In this way we avoid having to call the command from Django code and we
will be able to better test the bulk describe feature.

https://github.com/MTG/freesound/pull/1093
